TUC-129 Huge cleaning of numbers conversions

+++ src/tucucommon/utils.h
@@ -101,7 +101,7 @@ std::string getAppFolder(char** _argv);
 template<typename ... Args>
 std::string strFormat(const std::string& _format, Args ... _args)
 {
-    size_t size = std::snprintf(nullptr, 0, _format.c_str(), _args ...) + 1; // Extra space for '\0'
+    size_t size = static_cast<size_t>(std::snprintf(nullptr, 0, _format.c_str(), _args ...) + 1); // Extra space for '\0'
     std::unique_ptr<char[]> buf(new char[size]);
     std::snprintf(buf.get(), size, _format.c_str(), _args ...);
     return std::string(buf.get(), buf.get() + size - 1);                    // We don't want the '\0' inside
